WIP REFACTOR Users and Account components to views folder
User login/register works in dev mode

TODO: Refactor user input html controls to their own .vue file
