+++ src/views/User/account/Register.cy.js
