+++ src/views/User/account/Register.vue
